+++ R/ForwardSearch.R
@@ -1,25 +1,24 @@
 ForwardSearch <- function(spect, sigs, m.opts, max.mc.cores, p.thresh) {
   start <- OptimizeExposure(spectrum = spect, sigs = sigs, m.opts = m.opts)
   lh.w.all <- start$loglh # The likelihood with all signatures
-
-  if (m.opts$trace >= 1) {
-    message("log likelihood using all signatures = ", lh.w.all)
-  }
-
   num_sig <- ncol(sigs)
+  
+  if (num_sig == 1) {
+    return(optimal.exposure = start$exposure)
+  }
 
   if (m.opts$trace >= 0) {
     message("\nStarting forward search for ", colnames(spect))
+  }
+  
+  if (m.opts$trace >= 1) {
+    message("log likelihood using all signatures = ", lh.w.all)
   }
 
   sigs.to.choose <- sigs
   optimal.sigs <- sigs[, 0, drop = FALSE]
 
-  if (ncol(sigs) == 1) {
-    return(optimal.exposure = start$exposure)
-  }
-
-  for (step in seq_len(ncol(sigs) - 1)) {
+  for (step in seq_len(num_sig - 1)) {
     if (m.opts$trace >= 0) {
       message(
         "\nForward search step ", step, ": ",
@@ -46,7 +45,7 @@ ForwardSearch <- function(spect, sigs, m.opts, max.mc.cores, p.thresh) {
         if (is.infinite(try.exp$loglh)) {
           # It was not possible to generate the spectrum from the signatures, so
           # try.exp$loglh should be -Inf and the p value the test that the spectrum
-          # can be better reconstructed with sigs than with optimal.sigs is 0.
+          # can be better reconstructed with optimal.sigs than with sigs is 0.
           if (try.exp$loglh < 0) {
             chisq.p <- 0
           } else {
@@ -98,7 +97,7 @@ ForwardSearch <- function(spect, sigs, m.opts, max.mc.cores, p.thresh) {
 
     # At the last step, if the largest.p.value is still <= p.thresh,
     # then we will use all signatures
-    if (step == ncol(sigs) - 1) {
+    if (step == num_sig - 1) {
       if (m.opts$trace >= 0) {
         message("\nNeed to use all signatures to reconstruct the spectrum: ", 
                 paste(colnames(sigs), collapse = " "))